pyproject.toml
[tool.poetry]
name = "test_tech_servier"
version = "0.1.0"
description = ""
authors = ["changyi <song.c@sfeir.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pyspark = "^3.5.5"
loguru = "^0.7.3"
python-dotenv = "^1.0.1"

[tool.poetry.dev-dependencies]
black = "^25.1.0"
pylint = "^3.3.5"
ruff = "^0.11.0"
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
builtins = ["_"]
line-length = 150
target-version = "py311"

[tool.ruff.lint]
extend-select = ["I"]

[tool.ruff.format]
docstring-code-format = true